06/09/25

+++ primeiro-projeto/projeto/src/main/java/entidades/Venda.java
@@ -17,7 +17,7 @@ public class Venda {
     public Integer codigo;
 
     public LocalDateTime horaVenda;
-    public Integer valorT;
+    public Double valorT;
     public Integer quantT;
 
     @ManyToOne
